split up the different types of endpoints in 3 separate API namespaces
